requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
pillow
psycopg2-binary
python-decouple
drf-yasg

